Move boot.lua to src/modules/love/ with raw string literal format.
Split it up into 3 files to keep under VS2013's raw string literal character limit.
Now boot.lua doesn't need a regeneration step when it's modified.
